requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
decorator==4.4.2
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.5
numpy==1.23.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proglog==0.1.12
Pygments==2.19.2
pytest==8.4.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scipy==1.15.3
sk-video==1.1.10
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
